# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Aster DNS API
  version: 1.0.0
  description: |
    REST API for the self-hosted Aster DNS control plane. DNS Providers are the
    authoritative source for zones and records. Cookie-authenticated mutations
    require both a trusted Origin and the X-CSRF-Token header.
servers:
  - url: /
...
paths:
...
  /api/v1/users:
    get:
      tags: [Users]
      summary: List users
      description: Admin only.
      operationId: listUsers
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Users
          content:
            application/json:
              schema:
                type: object
                required: [users]
                properties:
                  users:
                    type: array
                    items: { $ref: '#/components/schemas/User' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Users]
      summary: Create a user and one-time enrollment token
      description: Admin only. The enrollment token is shown once.
      operationId: createUser
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CSRFToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUser' }
      responses:
        '201':
          description: User and one-time enrollment token
          content:
            application/json:
              schema:
                type: object
                required: [user, enrollment_token, enrollment_expires_at]
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  enrollment_token: { type: string, description: Returned once. }
                  enrollment_expires_at: { type: string, format: date-time }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/users/{id}:
    patch:
      tags: [Users]
      summary: Update a user
      description: Admin only.
      operationId: updateUser
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/ResourceID'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                display_name: { type: string }
                email: { type: string, format: email }
                role: { $ref: '#/components/schemas/Role' }
                password: { type: string, writeOnly: true }
                password_enabled: { type: boolean }
      responses:
        '200': { $ref: '#/components/responses/User' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/users/{id}/disable:
    post:
      tags: [Users]
      summary: Disable a user
      operationId: disableUser
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/ResourceID'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody: { $ref: '#/components/requestBodies/EmptyJSON' }
      responses:
        '200': { $ref: '#/components/responses/User' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/users/{id}/enable:
    post:
      tags: [Users]
      summary: Enable a user
      operationId: enableUser
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/ResourceID'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody: { $ref: '#/components/requestBodies/EmptyJSON' }
      responses:
        '200': { $ref: '#/components/responses/User' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/users/{id}/enrollment-token:
    post:
      tags: [Users]
      summary: Issue a replacement one-time enrollment token
      operationId: issueEnrollmentToken
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/ResourceID'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody: { $ref: '#/components/requestBodies/EmptyJSON' }
      responses:
        '200':
          description: One-time enrollment token
          content:
            application/json:
              schema:
                type: object
                required: [enrollment_token, enrollment_expires_at]
                properties:
                  enrollment_token: { type: string, description: Returned once. }
                  enrollment_expires_at: { type: string, format: date-time }
        default: { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-aster_session
      description: Opaque session cookie. HTTP deployments use the non-Host-prefixed equivalent.
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
      description: Must match the readable CSRF cookie for the current server-side session.
    ResourceID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  requestBodies:
    EmptyJSON:
      required: false
      content:
        application/json:
          schema: { type: object, maxProperties: 0 }
  responses:
    Error:
      description: Stable safe error. Details never contain Provider secrets or raw upstream payloads.
      headers:
        X-Request-ID:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    User:
      description: Safe user state
      content:
        application/json:
          schema:
            type: object
            required: [user]
            properties:
              user: { $ref: '#/components/schemas/User' }
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ItemError' }
    ItemError:
      type: object
      required: [code, message, request_id]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
        details: { type: object, additionalProperties: true }
...
    Role:
      type: string
      enum: [admin, operator, viewer]
    User:
      type: object
      required: [id, username, display_name, role, password_enabled, totp_required, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        display_name: { type: string }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/Role' }
        password_enabled: { type: boolean }
        totp_required: { type: boolean }
        disabled_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateUser:
      type: object
      additionalProperties: false
      required: [username, display_name, role]
      properties:
        username: { type: string }
        display_name: { type: string }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/Role' }
        initial_password: { type: string, minLength: 12, maxLength: 1024, writeOnly: true }
...